third_party/WebKit/LayoutTests/fast/animation/request-animation-frame-cancel2-expected.txt

Tests one requestAnimationFrame callback cancelling a second

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS successfullyParsed is true

TEST COMPLETE

PASS callbackFired is false
PASS cancelFired is true
PASS successfullyParsed is true

TEST COMPLETE
